+++ src/main/java/fi/juhpaza/staffactivity/gui/StaffActivityGuiView.java
@@ -4,6 +4,10 @@ package fi.juhpaza.staffactivity.gui;
  * Identifies which StaffActivity GUI view is open.
  */
 public enum StaffActivityGuiView {
-    ADMIN,
+    DASHBOARD,
+    ONLINE_STAFF,
+    TOP_STATISTICS,
+    TODAY_STATISTICS,
+    WEEKLY_STATISTICS,
     STAFF_SUMMARY
 }